requirements.txt
PyQt5
qtawesome
pandas
openpyxl
pyinstaller
pillow
python-docx